separated fixtures in more files added test to check dummmy service validity
